bardie.js: update: baseurl
Every time the URL changes anywhere in the app, it needs to be changed in multiple places, ESPECIALLY in bardie.js, because the package won't work otherwise.

This resolves two things - it references the URL in the package.json and is all maintained in one place, and it fixes the package!

Bardie now works again.