bardie.js: update: functionality, JSdocs
This commit merges the functions into one.

The only difference between the two was the optional "image" property.

Now, the code has been refactored so only one function is needed and "image" can be defined optionally.

This commit also includes a fix for importing the baseurl from the package.json, as well as making the JSDocs more concise,